application/python_scripts/requirements.txt
autopep8==1.7.0
coloredlogs==15.0.1
contourpy==1.0.5
cycler==0.11.0
ffmpeg-python==0.2.0
flatbuffers==22.11.23
fonttools==4.37.4
future==0.18.2
humanfriendly==10.0
kiwisolver==1.4.4
matplotlib==3.6.1
mpmath==1.2.1
numpy==1.23.3
onnx==1.12.0
onnxruntime==1.13.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.2
Pillow==9.2.0
protobuf==3.20.1
PyAudio==0.2.13
pycodestyle==2.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
# pyvirtualcam==0.10.0
PyYAML==6.0
six==1.16.0
sympy==1.11.1
toml==0.10.2
typing_extensions==4.4.0
